Refactor events models to separate files
