+++ docs/events.models.rst
@@ -1,0 +1,43 @@
+events.models package
+=====================
+
+.. automodule:: events.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+events.models.event module
+--------------------------
+
+.. automodule:: events.models.event
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.models.event\_registration module
+----------------------------------------
+
+.. automodule:: events.models.event_registration
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.models.feed\_token module
+--------------------------------
+
+.. automodule:: events.models.feed_token
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+events.models.registration\_information\_field module
+-----------------------------------------------------
+
+.. automodule:: events.models.registration_information_field
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
