+++ website/events/models/__init__.py
@@ -1,0 +1,6 @@
+"""The models defined by the events package"""
+
+from .event import *
+from .event_registration import *
+from .feed_token import *
+from .registration_information_field import *
